Split ProgrammingSubmissionService into multiple smaller service to avoid dependency hell
