+++ src/main/java/de/tum/in/www1/artemis/service/programming/ProgrammingExerciseService.java
@@ -184,7 +184,7 @@ public class ProgrammingExerciseService {
         programmingExerciseTaskService.updateTasksFromProblemStatement(programmingExercise);
 
         // The creation of the webhooks must occur after the initial push, because the participation is
-        // not yet saved in the database, so we cannot save the submission accordingly (see ProgrammingSubmissionService.notifyPush)
+        // not yet saved in the database, so we cannot save the submission accordingly (see ProgrammingSubmissionService.processNewProgrammingSubmission)
         versionControlService.get().addWebHooksForExercise(programmingExercise);
         scheduleOperations(programmingExercise.getId());
         groupNotificationService.checkNotificationsForNewExercise(programmingExercise);
