+++ src/main/java/de/tum/in/www1/artemis/service/programming/ProgrammingSubmissionService.java
@@ -11,11 +11,7 @@ import javax.annotation.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.actuate.audit.AuditEvent;
-import org.springframework.boot.actuate.audit.AuditEventRepository;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.messaging.simp.SimpMessageSendingOperations;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
 import de.tum.in.www1.artemis.domain.*;
@@ -33,10 +29,8 @@ import de.tum.in.www1.artemis.service.connectors.VersionControlService;
 import de.tum.in.www1.artemis.service.exam.ExamDateService;
 import de.tum.in.www1.artemis.service.exam.ExamSubmissionService;
 import de.tum.in.www1.artemis.service.hestia.ProgrammingExerciseGitDiffReportService;
-import de.tum.in.www1.artemis.service.notifications.GroupNotificationService;
 import de.tum.in.www1.artemis.web.rest.errors.AccessForbiddenException;
 import de.tum.in.www1.artemis.web.rest.errors.EntityNotFoundException;
-import de.tum.in.www1.artemis.web.websocket.programmingSubmission.BuildTriggerWebsocketError;
 
 @Service
 public class ProgrammingSubmissionService extends SubmissionService {
@@ -49,15 +43,11 @@ public class ProgrammingSubmissionService extends SubmissionService {
     @Value("${artemis.git.email}")
     private String artemisGitEmail;
 
-    @Value("${artemis.external-system-request.batch-size}")
-    private int externalSystemRequestBatchSize;
-
-    @Value("${artemis.external-system-request.batch-waiting-time}")
-    private int externalSystemRequestBatchWaitingTime;
-
     private final ProgrammingExerciseRepository programmingExerciseRepository;
 
     private final ProgrammingSubmissionRepository programmingSubmissionRepository;
+
+    private final ProgrammingMessagingService programmingMessagingService;
 
     private final ResultRepository resultRepository;
 
@@ -65,47 +55,36 @@ public class ProgrammingSubmissionService extends SubmissionService {
 
     private final ExamSubmissionService examSubmissionService;
 
-    private final GroupNotificationService groupNotificationService;
-
-    private final WebsocketMessagingService websocketMessagingService;
-
     private final Optional<VersionControlService> versionControlService;
 
     private final Optional<ContinuousIntegrationService> continuousIntegrationService;
 
     private final GitService gitService;
 
-    private final SimpMessageSendingOperations messagingTemplate;
-
-    private final AuditEventRepository auditEventRepository;
-
     private final ProgrammingExerciseStudentParticipationRepository programmingExerciseStudentParticipationRepository;
 
     private final ProgrammingExerciseGitDiffReportService programmingExerciseGitDiffReportService;
 
     public ProgrammingSubmissionService(ProgrammingSubmissionRepository programmingSubmissionRepository, ProgrammingExerciseRepository programmingExerciseRepository,
-            GroupNotificationService groupNotificationService, SubmissionRepository submissionRepository, UserRepository userRepository, AuthorizationCheckService authCheckService,
-            WebsocketMessagingService websocketMessagingService, Optional<VersionControlService> versionControlService, ResultRepository resultRepository,
-            Optional<ContinuousIntegrationService> continuousIntegrationService, ParticipationService participationService, SimpMessageSendingOperations messagingTemplate,
+            SubmissionRepository submissionRepository, UserRepository userRepository, AuthorizationCheckService authCheckService,
+            ProgrammingMessagingService programmingMessagingService, Optional<VersionControlService> versionControlService, ResultRepository resultRepository,
+            Optional<ContinuousIntegrationService> continuousIntegrationService, ParticipationService participationService,
             ProgrammingExerciseParticipationService programmingExerciseParticipationService, ExamSubmissionService examSubmissionService, GitService gitService,
-            StudentParticipationRepository studentParticipationRepository, FeedbackRepository feedbackRepository, AuditEventRepository auditEventRepository,
-            ExamDateService examDateService, ExerciseDateService exerciseDateService, CourseRepository courseRepository, ParticipationRepository participationRepository,
+            StudentParticipationRepository studentParticipationRepository, FeedbackRepository feedbackRepository, ExamDateService examDateService,
+            ExerciseDateService exerciseDateService, CourseRepository courseRepository, ParticipationRepository participationRepository,
             ProgrammingExerciseStudentParticipationRepository programmingExerciseStudentParticipationRepository, ComplaintRepository complaintRepository,
             ProgrammingExerciseGitDiffReportService programmingExerciseGitDiffReportService) {
         super(submissionRepository, userRepository, authCheckService, resultRepository, studentParticipationRepository, participationService, feedbackRepository, examDateService,
                 exerciseDateService, courseRepository, participationRepository, complaintRepository);
         this.programmingSubmissionRepository = programmingSubmissionRepository;
         this.programmingExerciseRepository = programmingExerciseRepository;
-        this.groupNotificationService = groupNotificationService;
-        this.websocketMessagingService = websocketMessagingService;
+        this.programmingMessagingService = programmingMessagingService;
         this.versionControlService = versionControlService;
         this.continuousIntegrationService = continuousIntegrationService;
-        this.messagingTemplate = messagingTemplate;
         this.programmingExerciseParticipationService = programmingExerciseParticipationService;
         this.examSubmissionService = examSubmissionService;
         this.gitService = gitService;
         this.resultRepository = resultRepository;
-        this.auditEventRepository = auditEventRepository;
         this.programmingExerciseStudentParticipationRepository = programmingExerciseStudentParticipationRepository;
         this.programmingExerciseGitDiffReportService = programmingExerciseGitDiffReportService;
     }
@@ -120,7 +99,8 @@ public class ProgrammingSubmissionService extends SubmissionService {
      * @throws IllegalStateException    if a ProgrammingSubmission already exists
      * @throws IllegalArgumentException if the Commit hash could not be parsed for submission from participation
      */
-    public ProgrammingSubmission notifyPush(Long participationId, Object requestBody) throws EntityNotFoundException, IllegalStateException, IllegalArgumentException {
+    public ProgrammingSubmission processNewProgrammingSubmission(Long participationId, Object requestBody)
+            throws EntityNotFoundException, IllegalStateException, IllegalArgumentException {
         // Note: the following line is intentionally at the top of the method to get the most accurate submission date
         ZonedDateTime submissionDate = ZonedDateTime.now();
         Participation participation = participationRepository.findByIdWithLegalSubmissionsElseThrow(participationId);
@@ -183,17 +163,13 @@ public class ProgrammingSubmissionService extends SubmissionService {
         log.info("Create new programmingSubmission with commitHash: {} for participation {}", commit.getCommitHash(), participationId);
 
         programmingSubmission.setSubmitted(true);
-
         programmingSubmission.setSubmissionDate(submissionDate);
         programmingSubmission.setType(SubmissionType.MANUAL);
 
         // Students are not allowed to submit a programming exercise after the exam due date, if this happens we set the Submission to ILLEGAL
         checkForIllegalExamSubmission(programmingExerciseParticipation, programmingSubmission);
-
         programmingExerciseParticipation.addSubmission(programmingSubmission);
-
         programmingSubmission = programmingSubmissionRepository.save(programmingSubmission);
-
         updateGitDiffReport(programmingExerciseParticipation);
 
         // NOTE: we don't need to save the participation here, this might lead to concurrency problems when doing the empty commit during resume exercise!
@@ -237,7 +213,7 @@ public class ProgrammingSubmissionService extends SubmissionService {
                         + " just illegally submitted code after the allowed individual due date (including the grace period) in the participation "
                         + programmingExerciseParticipation.getId() + " for the exam programming exercise " + programmingExercise.getId();
                 programmingSubmission.setType(SubmissionType.ILLEGAL);
-                groupNotificationService.notifyInstructorGroupAboutIllegalSubmissionsForExercise(programmingExercise, message);
+                programmingMessagingService.notifyInstructorGroupAboutIllegalSubmissionsForExercise(programmingExercise, message);
                 log.warn(message);
             }
         }
@@ -294,80 +270,11 @@ public class ProgrammingSubmissionService extends SubmissionService {
     }
 
     /**
-     * Trigger the CI of all student participations and the template participation of the given exercise.
-     * The build result will become rated regardless of the due date as the submission type is INSTRUCTOR.
-     *
-     * The method is async because it would time out a calling resource method.
-     *
-     * @param exerciseId to identify the programming exercise.
-     * @throws EntityNotFoundException if there is no programming exercise for the given exercise id.
-     */
-    @Async
-    public void triggerInstructorBuildForExercise(Long exerciseId) throws EntityNotFoundException {
-        // Async can't access the authentication object. We need to do any security checks before this point.
-        SecurityUtils.setAuthorizationObject();
-        var programmingExercise = programmingExerciseRepository.findByIdElseThrow(exerciseId);
-
-        // Let the instructor know that a build run was triggered.
-        notifyInstructorAboutStartedExerciseBuildRun(programmingExercise);
-        List<ProgrammingExerciseStudentParticipation> participations = new ArrayList<>(
-                programmingExerciseStudentParticipationRepository.findWithSubmissionsByExerciseId(exerciseId));
-
-        triggerBuildForParticipations(participations);
-
-        // When the instructor build was triggered for the programming exercise, it is not considered 'dirty' anymore.
-        setTestCasesChanged(programmingExercise, false);
-        // Let the instructor know that the build run is finished.
-        notifyInstructorAboutCompletedExerciseBuildRun(programmingExercise);
-    }
-
-    /**
-     * trigger the build using the batch size approach for all participations
-     *
-     * @param participations the participations for which the method triggerBuild should be executed.
-     */
-    public void triggerBuildForParticipations(List<ProgrammingExerciseStudentParticipation> participations) {
-        var index = 0;
-        for (var participation : participations) {
-            // Execute requests in batches instead all at once.
-            if (index > 0 && index % externalSystemRequestBatchSize == 0) {
-                try {
-                    log.info("Sleep for {}s during triggerBuild", externalSystemRequestBatchWaitingTime / 1000);
-                    Thread.sleep(externalSystemRequestBatchWaitingTime);
-                }
-                catch (InterruptedException ex) {
-                    log.error("Exception encountered when pausing before executing successive build for participation {}", participation.getId(), ex);
-                }
-            }
-            triggerBuild(participation);
-            index++;
-        }
-    }
-
-    public void logTriggerInstructorBuild(User user, Exercise exercise, Course course) {
-        var auditEvent = new AuditEvent(user.getLogin(), TRIGGER_INSTRUCTOR_BUILD, "exercise=" + exercise.getTitle(), "course=" + course.getTitle());
-        auditEventRepository.add(auditEvent);
-        log.info("User {} triggered an instructor build for all participations in exercise {} with id {}", user.getLogin(), exercise.getTitle(), exercise.getId());
-    }
-
-    private void notifyInstructorAboutStartedExerciseBuildRun(ProgrammingExercise programmingExercise) {
-        websocketMessagingService.sendMessage(getProgrammingExerciseAllExerciseBuildsTriggeredTopic(programmingExercise.getId()), BuildRunState.RUNNING);
-        // Send a notification to the client to inform the instructor about the test case update.
-        groupNotificationService.notifyEditorAndInstructorGroupAboutExerciseUpdate(programmingExercise, BUILD_RUN_STARTED_FOR_PROGRAMMING_EXERCISE);
-    }
-
-    private void notifyInstructorAboutCompletedExerciseBuildRun(ProgrammingExercise programmingExercise) {
-        websocketMessagingService.sendMessage(getProgrammingExerciseAllExerciseBuildsTriggeredTopic(programmingExercise.getId()), BuildRunState.COMPLETED);
-        // Send a notification to the client to inform the instructor about the test case update.
-        groupNotificationService.notifyEditorAndInstructorGroupAboutExerciseUpdate(programmingExercise, BUILD_RUN_COMPLETE_FOR_PROGRAMMING_EXERCISE);
-    }
-
-    /**
      * Create a submission with given submission type for the last commit hash of the given participation.
      * WARNING: The commitHash is used to map incoming results to submissions. Using this method could cause the result to have multiple fitting submissions.
-     *
-     * See discussion in: https://github.com/ls1intum/Artemis/pull/712#discussion_r314944129;
-     *
+     * <p>
+     * See <a href="https://github.com/ls1intum/Artemis/pull/712#discussion_r314944129">discussion</a>
+     * <p>
      * Worst case scenario when using this method:
      * 1) Student executes a submission, the build is created on Bamboo
      * 2) The build takes longer than 2 minutes, this enables the student to trigger the submission again
@@ -435,218 +342,16 @@ public class ProgrammingSubmissionService extends SubmissionService {
     }
 
     /**
-     * Trigger a CI build for each submission & notify each user of the participation
-     * Note: Instead of triggering all builds at the same time, we execute the builds in batches to not overload the CIS system (this has to be handled in the invoking method)
-     *
-     * Note: This call "resumes the exercise", i.e. re-creates the build plan if the build plan was already cleaned before
-     *
-     * @param participation the participation for which we create a new submission and new result
-     */
-    public void triggerBuild(ProgrammingExerciseStudentParticipation participation) {
-        Optional<ProgrammingSubmission> submission = participation.findLatestSubmission();
-        // we only need to trigger the build if the student actually already made a submission, otherwise this is not needed
-        if (submission.isPresent()) {
-            try {
-                if (participation.getBuildPlanId() == null || !participation.getInitializationState().hasCompletedState(InitializationState.INITIALIZED)) {
-                    // in this case, we first have to resume the exercise: this includes that we again set up the build plan properly before we trigger it
-                    participationService.resumeProgrammingExercise(participation);
-                    // Note: in this case we do not need an empty commit: when we trigger the build manually (below), subsequent commits will work correctly
-                }
-                continuousIntegrationService.get().triggerBuild(participation);
-                // TODO: this is a workaround, in the future we should use the participation to notify the client and avoid using the submission
-                this.notifyUserAboutSubmission(submission.get());
-            }
-            catch (Exception e) {
-                log.error("Trigger build failed for {} with the exception {}", participation.getBuildPlanId(), e.getMessage());
-                BuildTriggerWebsocketError error = new BuildTriggerWebsocketError(e.getMessage(), participation.getId());
-                notifyUserAboutSubmissionError(participation, error);
-            }
-        }
-    }
-
-    /**
-     * Triggers a build on the CI system and sends a websocket message to the user about the new submission and
-     * Will send an error object in the case that the communication with the CI failed.
-     *
-     * Note: This call "resumes the exercise", i.e. re-creates the build plan if the build plan was already cleaned before
-     *
-     * @param submission ProgrammingSubmission that was just created.
-     */
-    public void triggerBuildAndNotifyUser(ProgrammingSubmission submission) {
-        var programmingExerciseParticipation = (ProgrammingExerciseParticipation) submission.getParticipation();
-        try {
-            if (programmingExerciseParticipation instanceof ProgrammingExerciseStudentParticipation && (programmingExerciseParticipation.getBuildPlanId() == null
-                    || !programmingExerciseParticipation.getInitializationState().hasCompletedState(InitializationState.INITIALIZED))) {
-
-                // in this case, we first have to resume the exercise: this includes that we again setup the build plan properly before we trigger it
-                participationService.resumeProgrammingExercise((ProgrammingExerciseStudentParticipation) programmingExerciseParticipation);
-                // Note: in this case we do not need an empty commit: when we trigger the build manually (below), subsequent commits will work correctly
-            }
-            continuousIntegrationService.get().triggerBuild(programmingExerciseParticipation);
-            notifyUserAboutSubmission(submission);
-        }
-        catch (Exception e) {
-            log.error("Trigger build failed for {} with the exception {}", programmingExerciseParticipation.getBuildPlanId(), e.getMessage());
-            BuildTriggerWebsocketError error = new BuildTriggerWebsocketError(e.getMessage(), submission.getParticipation().getId());
-            notifyUserAboutSubmissionError(submission, error);
-        }
-    }
-
-    /**
-     * Trigger the template repository build with the given commitHash.
-     *
-     * @param programmingExerciseId is used to retrieve the template participation.
-     * @param commitHash            the unique hash code of the git repository identifying the submission, will be used for the created submission.
-     * @param submissionType        will be used for the created submission.
-     * @throws EntityNotFoundException if the programming exercise has no template participation (edge case).
-     */
-    public void triggerTemplateBuildAndNotifyUser(long programmingExerciseId, String commitHash, SubmissionType submissionType) throws EntityNotFoundException {
-        TemplateProgrammingExerciseParticipation templateParticipation = programmingExerciseParticipationService
-                .findTemplateParticipationByProgrammingExerciseId(programmingExerciseId);
-        // If for some reason the programming exercise does not have a template participation, we can only log and abort.
-        createSubmissionTriggerBuildAndNotifyUser(templateParticipation, commitHash, submissionType);
-    }
-
-    /**
-     * Creates a submission with the given type and commitHash for the provided participation.
-     * Will notify the user about occurring errors when trying to trigger the build.
-     *
-     * @param participation  for which to create the submission.
-     * @param commitHash     the unique hash code of the git repository identifying the submission,to assign to the submission.
-     * @param submissionType to assign to the submission.
-     */
-    private void createSubmissionTriggerBuildAndNotifyUser(ProgrammingExerciseParticipation participation, String commitHash, SubmissionType submissionType) {
-        ProgrammingSubmission submission = createSubmissionWithCommitHashAndSubmissionType(participation, commitHash, submissionType);
-        try {
-            continuousIntegrationService.get().triggerBuild((ProgrammingExerciseParticipation) submission.getParticipation());
-            notifyUserAboutSubmission(submission);
-        }
-        catch (Exception e) {
-            BuildTriggerWebsocketError error = new BuildTriggerWebsocketError(e.getMessage(), submission.getParticipation().getId());
-            notifyUserAboutSubmissionError(submission, error);
-        }
-    }
-
-    /**
-     * Executes setTestCasesChanged with testCasesChanged = true, also triggers template and solution build.
-     * This method should be used if the solution participation would otherwise not be built.
-     *
-     * @param programmingExerciseId ProgrammingExercise id
-     * @throws EntityNotFoundException if there is no programming exercise for the given id.
-     */
-    public void setTestCasesChangedAndTriggerTestCaseUpdate(long programmingExerciseId) throws EntityNotFoundException {
-        setTestCasesChanged(programmingExerciseId, true);
-        var programmingExercise = programmingExerciseRepository.findWithTemplateAndSolutionParticipationById(programmingExerciseId).get();
-
-        try {
-            continuousIntegrationService.get().triggerBuild(programmingExercise.getSolutionParticipation());
-            continuousIntegrationService.get().triggerBuild(programmingExercise.getTemplateParticipation());
-        }
-        catch (ContinuousIntegrationException ex) {
-            log.error("Could not trigger build for solution repository after test case update for programming exercise with id {}", programmingExerciseId);
-        }
-    }
-
-    /**
-     * see the description below
-     *
-     * @param programmingExerciseId id of a ProgrammingExercise.
-     * @param testCasesChanged      set to true to mark the programming exercise as dirty.
-     * @throws EntityNotFoundException if the programming exercise does not exist.
-     */
-    public void setTestCasesChanged(long programmingExerciseId, boolean testCasesChanged) throws EntityNotFoundException {
-        var programmingExercise = programmingExerciseRepository.findByIdElseThrow(programmingExerciseId);
-        setTestCasesChanged(programmingExercise, testCasesChanged);
-    }
-
-    /**
-     * If testCasesChanged = true, this marks the programming exercise as dirty, meaning that its test cases were changed and the student submissions should be built & tested.
-     * This method also sends out a notification to the client if testCasesChanged = true.
-     * In case the testCaseChanged value is the same for the programming exercise or the programming exercise is not released or has no results, the method will return immediately.
-     *
-     * @param programmingExercise a ProgrammingExercise.
-     * @param testCasesChanged    set to true to mark the programming exercise as dirty.
-     * @throws EntityNotFoundException if the programming exercise does not exist.
-     */
-    public void setTestCasesChanged(ProgrammingExercise programmingExercise, boolean testCasesChanged) throws EntityNotFoundException {
-
-        // If the flag testCasesChanged has not changed, we can stop the execution
-        // Also, if the programming exercise has no results yet, there is no point in setting test cases changed to *true*.
-        // It is only relevant when there are student submissions that should get an updated result.
-
-        boolean resultsExist = resultRepository.existsByParticipation_ExerciseId(programmingExercise.getId());
-
-        if (testCasesChanged == programmingExercise.getTestCasesChanged() || (!resultsExist && testCasesChanged)) {
-            return;
-        }
-        programmingExercise.setTestCasesChanged(testCasesChanged);
-        ProgrammingExercise updatedProgrammingExercise = programmingExerciseRepository.save(programmingExercise);
-        // Send a websocket message about the new state to the client.
-        websocketMessagingService.sendMessage(getProgrammingExerciseTestCaseChangedTopic(updatedProgrammingExercise.getId()), testCasesChanged);
-        // Send a notification to the client to inform the instructor about the test case update.
-        if (testCasesChanged) {
-            groupNotificationService.notifyEditorAndInstructorGroupsAboutChangedTestCasesForProgrammingExercise(updatedProgrammingExercise);
-        }
-        else {
-            groupNotificationService.notifyEditorAndInstructorGroupAboutExerciseUpdate(updatedProgrammingExercise, TEST_CASES_CHANGED_RUN_COMPLETED_NOTIFICATION);
-        }
-    }
-
-    private String getProgrammingExerciseTestCaseChangedTopic(Long programmingExerciseId) {
-        return "/topic/programming-exercises/" + programmingExerciseId + "/test-cases-changed";
-    }
-
-    private String getProgrammingExerciseAllExerciseBuildsTriggeredTopic(Long programmingExerciseId) {
-        return "/topic/programming-exercises/" + programmingExerciseId + "/all-builds-triggered";
-    }
-
-    /**
-     * Notify user on a new programming submission.
-     *
-     * @param submission ProgrammingSubmission
-     */
-    public void notifyUserAboutSubmission(ProgrammingSubmission submission) {
-        if (submission.getParticipation() instanceof StudentParticipation studentParticipation) {
-            // no need to send all exercise details here
-            submission.getParticipation().setExercise(null);
-            studentParticipation.getStudents().forEach(user -> messagingTemplate.convertAndSendToUser(user.getLogin(), NEW_SUBMISSION_TOPIC, submission));
-        }
-
-        if (submission.getParticipation() != null && submission.getParticipation().getExercise() != null) {
-            var topicDestination = getExerciseTopicForTAAndAbove(submission.getParticipation().getExercise().getId());
-            messagingTemplate.convertAndSend(topicDestination, submission);
-        }
-    }
-
-    private void notifyUserAboutSubmissionError(ProgrammingSubmission submission, BuildTriggerWebsocketError error) {
-        notifyUserAboutSubmissionError(submission.getParticipation(), error);
-    }
-
-    private void notifyUserAboutSubmissionError(Participation participation, BuildTriggerWebsocketError error) {
-        if (participation instanceof StudentParticipation studentParticipation) {
-            studentParticipation.getStudents().forEach(user -> messagingTemplate.convertAndSendToUser(user.getLogin(), NEW_SUBMISSION_TOPIC, error));
-        }
-
-        if (participation != null && participation.getExercise() != null) {
-            messagingTemplate.convertAndSend(getExerciseTopicForTAAndAbove(participation.getExercise().getId()), error);
-        }
-    }
-
-    private String getExerciseTopicForTAAndAbove(long exerciseId) {
-        return EXERCISE_TOPIC_ROOT + exerciseId + PROGRAMMING_SUBMISSION_TOPIC;
-    }
-
-    /**
      * Given an exercise id and a tutor id, it returns all the programming submissions where the tutor has assessed a result
-     *
+     * <p>
      * Exam mode:
      * The query that returns the participations returns the contained submissions in a particular way:
      * Due to hibernate, the result list has null values at all places where a result was assessed by another tutor.
      * At the beginning of the list remain all the automatic results without assessor.
-     * Then filtering out all submissions which have no result at the place of the correctionround leaves us with the submissions we are interested in.
+     * Then filtering out all submissions which have no result at the place of the correction round leaves us with the submissions we are interested in.
      * Before returning the submissions we strip away all automatic results, to be able to correctly display them in the client.
-     *
-     * Not exam mode:
+     * <p>
+     * Not exam mode (i.e. course exercise):
      * In this case the query that returns the participations returns the contained submissions in a different way:
      * Here hibernate sets all automatic results to null, therefore we must filter all those out. This way the client can access the submissions'
      * single result.
@@ -661,14 +366,14 @@ public class ProgrammingSubmissionService extends SubmissionService {
             int correctionRound) {
         List<Submission> submissions;
         if (examMode) {
-            var participations = this.studentParticipationRepository.findAllByParticipationExerciseIdAndResultAssessorAndCorrectionRoundIgnoreTestRuns(exerciseId, tutor);
+            var participations = studentParticipationRepository.findAllByParticipationExerciseIdAndResultAssessorAndCorrectionRoundIgnoreTestRuns(exerciseId, tutor);
             // Latest submission might be illegal
             submissions = participations.stream().map(StudentParticipation::findLatestLegalOrIllegalSubmission).filter(Optional::isPresent).map(Optional::get)
                     // filter out the submissions that don't have a result (but a null value) for the correctionRound
                     .filter(submission -> submission.hasResultForCorrectionRound(correctionRound)).toList();
         }
         else {
-            submissions = this.submissionRepository.findAllByParticipationExerciseIdAndResultAssessor(exerciseId, tutor);
+            submissions = submissionRepository.findAllByParticipationExerciseIdAndResultAssessor(exerciseId, tutor);
             // automatic results are null in the received results list. We need to filter them out for the client to display the dashboard correctly
             submissions.forEach(Submission::removeNullResults);
         }
@@ -717,7 +422,7 @@ public class ProgrammingSubmissionService extends SubmissionService {
      * stored in the attribute participation.submissionCount instead of being determined by the size of the set of all submissions.
      * This method is intended to reduce the amount of data transferred to the client.
      * @param programmingSubmissionList - a List with all ProgrammingSubmissions to be modified
-     * @param removeSubmissionSet - option to also remove the SubmissionSet from the ProgrammingSubmssion
+     * @param removeSubmissionSet - option to also remove the SubmissionSet from the ProgrammingSubmission
      * @return a List with ProgrammingSubmissions and removed attributes
      */
     private List<ProgrammingSubmission> removeExerciseAndSubmissionSet(List<ProgrammingSubmission> programmingSubmissionList, boolean removeSubmissionSet) {
@@ -762,7 +467,7 @@ public class ProgrammingSubmissionService extends SubmissionService {
      * @return the locked programming submission
      */
     public ProgrammingSubmission lockAndGetProgrammingSubmission(Long submissionId, int correctionRound) {
-        ProgrammingSubmission programmingSubmission = findOneWithEagerResultsFeedbacksAssessor(submissionId);
+        ProgrammingSubmission programmingSubmission = programmingSubmissionRepository.findByIdWithEagerResultsFeedbacksAssessorElseThrow(submissionId);
         var manualResult = lockSubmission(programmingSubmission, correctionRound);
         return (ProgrammingSubmission) manualResult.getSubmission();
     }
@@ -837,10 +542,5 @@ public class ProgrammingSubmissionService extends SubmissionService {
         // Make sure that submission is set back after saving
         newResult.setSubmission(existingSubmission);
         return newResult;
-    }
-
-    private ProgrammingSubmission findOneWithEagerResultsFeedbacksAssessor(Long submissionId) {
-        return programmingSubmissionRepository.findWithEagerResultsFeedbacksAssessorById(submissionId)
-                .orElseThrow(() -> new EntityNotFoundException("Programming submission with id \"" + submissionId + "\" does not exist"));
     }
 }
